finance_api/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djoser>=2.2
psycopg2-binary>=2.9
python-dotenv>=1.0
